+++ osa4/blogilista/tests/blogs.test.js
@@ -46,15 +46,33 @@ const newBlog = {
     likes: 2
 };
 
+// Pakollisen autentikoinnin lisäämisen jälkeen
+// korjasin osan testeistä kirjautumista, mutta
+// en kaikkia, koska testien korjaamista
+// ei vaadittu tehtävissä.
+const login = {
+    username: "TestUser",
+    password: "password"
+};
+
 const app = require("../server.js");
 const Blog = require("../models/blogi.js").Blog;
+const User = require("../models/user.js").User;
 
 beforeEach(async () => {
     await Blog.deleteMany({});
+    await User.deleteMany({});
 
     for(let blog of initialBlogs){
         await new Blog(blog).save();
     }
+
+    await new User({
+        username: "TestUser",
+        password: "b109f3bbbc244eb82441917ed06d618b9008dd09b3befd1b5e07394c706a8bb980b1d7785e5976ec049b46df5f1326af5a2ea6d103fd07c95385ffab0cacbc86", // "password"
+        name: "Test user",
+        blogs: []
+    }).save();
 });
 
 // ### Actual tests ###
@@ -76,19 +94,22 @@ describe("GET /api/blogs",() => {
 
 describe("POST /api/blogs",() => {
     test("API should return the right number of blogs after POST",async () => {
-        await api.post("/api/blogs").send(newBlog).expect(201);
+        const loginResponse = await api.post("/api/login").send(login);
+        await api.post("/api/blogs").set("Authorization",`Bearer ${loginResponse.body.token}`).send(newBlog).expect(201);
         const response = await api.get("/api/blogs").expect(200).expect("Content-Type",/application\/json/);
         expect(response.body.length).toBe(initialBlogs.length + 1);
     });
 
     test("New blog should be correctly defined",async () => {
-        await api.post("/api/blogs").send(newBlog).expect(201);
+        const loginResponse = await api.post("/api/login").send(login);
+        await api.post("/api/blogs").set("Authorization",`Bearer ${loginResponse.body.token}`).send(newBlog).expect(201);
         const response = await api.get("/api/blogs").expect(200).expect("Content-Type",/application\/json/);
         expect(response.body[response.body.length - 1]).toMatchObject(newBlog);
     });
 
     test("All blogs should have property 'id'",async () => {
-        await api.post("/api/blogs").send(newBlog).expect(201);
+        const loginResponse = await api.post("/api/login").send(login);
+        await api.post("/api/blogs").set("Authorization",`Bearer ${loginResponse.body.token}`).send(newBlog).expect(201);
         const response = await api.get("/api/blogs").expect(200).expect("Content-Type",/application\/json/);
         response.body.forEach((blog) => {
             expect(blog.id).toBeDefined();
@@ -96,22 +117,26 @@ describe("POST /api/blogs",() => {
     });
 
     test("If number of likes is not given, likes should be set to zero",async () => {
+        const loginResponse = await api.post("/api/login").send(login);
+
         let copy = {...newBlog};
         delete copy.likes;
-        await api.post("/api/blogs").send(copy).expect(201);
+        await api.post("/api/blogs").set("Authorization",`Bearer ${loginResponse.body.token}`).send(copy).expect(201);
 
         const response = await api.get("/api/blogs").expect(200).expect("Content-Type",/application\/json/);
         expect(response.body[response.body.length - 1].likes).toBe(0);
     });
 
     test("New blogs without title or url should not be accepted",async () => {
+        const loginResponse = await api.post("/api/login").send(login);
+
         let copy = {...newBlog};
         delete copy.title;
-        await api.post("/api/blogs").send(copy).expect(400);
+        await api.post("/api/blogs").set("Authorization",`Bearer ${loginResponse.body.token}`).send(copy).expect(400);
 
         copy = {...newBlog};
         delete copy.url;
-        await api.post("/api/blogs").send(copy).expect(400);
+        await api.post("/api/blogs").set("Authorization",`Bearer ${loginResponse.body.token}`).send(copy).expect(400);
 
         const response = await api.get("/api/blogs").expect(200).expect("Content-Type",/application\/json/);
         expect(response.body.length).toBe(initialBlogs.length);
@@ -120,17 +145,21 @@ describe("POST /api/blogs",() => {
 
 describe("DELETE /api/blogs",() => {
     test("API should return the right number of blogs after DELETE",async () => {
+        const loginResponse = await api.post("/api/login").send(login);
+
         let id = await api.get("/api/blogs");
         id = id.body[0].id;
-        await api.delete(`/api/blogs/${id}`).expect(205);
+        await api.delete(`/api/blogs/${id}`).set("Authorization",`Bearer ${loginResponse.body.token}`).expect(205);
 
         const response = await api.get("/api/blogs").expect(200).expect("Content-Type",/application\/json/);
         expect(response.body.length).toBe(initialBlogs.length - 1);
     });
 
     test("Wrong or missing ID returns 404",async () => {
-        await api.delete("/api/blogs/LoremIpsum").expect(404);
-        await api.delete("/api/blogs/").expect(404);
+        const loginResponse = await api.post("/api/login").send(login);
+
+        await api.delete("/api/blogs/LoremIpsum").set("Authorization",`Bearer ${loginResponse.body.token}`).expect(404);
+        await api.delete("/api/blogs/").set("Authorization",`Bearer ${loginResponse.body.token}`).expect(404);
 
         const response = await api.get("/api/blogs").expect(200).expect("Content-Type",/application\/json/);
         expect(response.body.length).toBe(initialBlogs.length);
